working app.ts
